Unit Test: MD5 txt files list update

+++ UnitTest/UnitTest.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Windows.Forms;
-using Cryptography;
-using System.IO;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using System.Configuration;
+using System.IO;
+using System.Linq;
+using Cryptography;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace UnitTest
 {
@@ -16,29 +12,46 @@ namespace UnitTest
     {
         [TestMethod]
         public void TestMd5TxtList()
-        {                        
-            var dir = ConfigurationManager.AppSettings.Get("test_dir");            
+        {
+            var dir = ConfigurationManager.AppSettings.Get("test_dir");
             var text_1 = ConfigurationManager.AppSettings.Get("txt_file_text_1");
             var md5_1 = ConfigurationManager.AppSettings.Get("txt_file_md5_1").ToUpper();
-
-            var list = new List<string>() { dir + "/test1.txt", dir + "/test2.txt", dir + "/test3.txt" };
-            var sol = new List<string>() { list.ElementAt(0) + " = " + md5_1, list.ElementAt(1) + " = " + md5_1, list.ElementAt(2) + " = " + md5_1 };
-
-            if (!Directory.Exists(dir))
-                Directory.CreateDirectory(dir);                
-
+            var text_2 = ConfigurationManager.AppSettings.Get("txt_file_text_2");
+            var md5_2 = ConfigurationManager.AppSettings.Get("txt_file_md5_2").ToUpper();
+            var text_3 = ConfigurationManager.AppSettings.Get("txt_file_text_3");
+            var md5_3 = ConfigurationManager.AppSettings.Get("txt_file_md5_3").ToUpper();
+            var list = new List<string> {dir + "/test1.txt", dir + "/test2.txt", dir + "/test3.txt"};
+            var sol = new List<string>
+            {
+                list.ElementAt(0) + " = " + md5_1,
+                list.ElementAt(1) + " = " + md5_2,
+                list.ElementAt(2) + " = " + md5_3
+            };
+            if (!Directory.Exists(dir)) Directory.CreateDirectory(dir);
             foreach (var l in list)
             {
-                if (File.Exists(l))
-                    File.Delete(l);
-                using (StreamWriter sw = new StreamWriter(l, true))
+                if (File.Exists(l)) File.Delete(l);
+                using (var sw = new StreamWriter(l, true))
                 {
-                    sw.WriteLine(text_1);
+                    switch (list.IndexOf(l))
+                    {
+                        case 0:
+                            sw.WriteLine(text_1);
+                            break;
+                        case 1:
+                            sw.WriteLine(text_2);
+                            break;
+                        case 2:
+                            sw.WriteLine(text_3);
+                            break;
+                        default: break;
+                    }
+
                     sw.Close();
                 }
             }
 
-            CollectionAssert.AreEqual(ClassAcm.CalculateTxtMd5List(list).ToList(), sol); 
+            CollectionAssert.AreEqual(ClassAcm.CalculateTxtMd5List(list).ToList(), sol);
         }
     }
 }
